+++ src/routes/web.php
@@ -1,7 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-use VeskoDigital\LaravelFCM\Http\Controllers\LaravelFcmController;
+use Nylo\LaravelFCM\Http\Controllers\LaravelFcmController;
 
 Route::put('device', [LaravelFcmController::class, 'update'])->name('laravel_notify_fcm.update');